added authentication

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -6,6 +6,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -47,11 +48,13 @@ public class AuctionsController : ControllerBase {
         return _mapper.Map<AuctionDto>(auction);
     }
 
+    [Authorize]
     [HttpPost]
     public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto){
         var auction = _mapper.Map<Auction>(auctionDto);
-        //TODO: add current user as seller
-        auction.Seller = "test";
+        
+        auction.Seller = User.Identity.Name;
+        Console.WriteLine("Username: " + User.Identity.Name);
 
         _context.Auctions.Add(auction);
 
@@ -66,6 +69,7 @@ public class AuctionsController : ControllerBase {
         return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, _mapper.Map<AuctionDto>(auction));
     }
 
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto){
         var auction = await _context.Auctions.Include(x => x.Item)
@@ -73,22 +77,13 @@ public class AuctionsController : ControllerBase {
 
         if (auction == null) return NotFound();
 
-        //TODO: check seller == username
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
         auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;   
         auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;   
         auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;   
         auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
-
-        var updatedData = new AuctionUpdated {
-            Year = auction.Item.Year,
-            Make = auction.Item.Make,
-            Model = auction.Item.Model,
-            Color = auction.Item.Color,
-            Mileage = auction.Item.Mileage,
-            Id = id.ToString()
-        };
 
         var updatedAuction = _mapper.Map<AuctionDto>(auction);
 
@@ -100,11 +95,14 @@ public class AuctionsController : ControllerBase {
         return BadRequest("Could not save changes");   
     }
 
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<ActionResult> DeleteAuction(Guid id){
         var auction = await _context.Auctions.FindAsync(id);
 
         if (auction == null) return NotFound();
+
+        if (auction.Seller != User.Identity.Name) return Forbid();
 
         _context.Auctions.Remove(auction);
